anthill structure and loottable chest
