+++ src/main/java/group/rys/common/item/ModMusicDiscItem.java
@@ -1,0 +1,11 @@
+package group.rys.common.item;
+
+import net.minecraft.item.Item;
+import net.minecraft.item.MusicDiscItem;
+import net.minecraft.util.SoundEvent;
+
+public class ModMusicDiscItem extends MusicDiscItem {
+    public ModMusicDiscItem(int comparatorValueIn, SoundEvent soundIn, Item.Properties builder) {
+        super(comparatorValueIn, soundIn, builder);
+    }
+}
