+++ src/main/java/group/rys/core/registry/other/ModProperties.java
@@ -29,5 +29,5 @@ public class ModProperties {
 	public static final Item.Properties apricot = new Item.Properties().group(Reference.MOD_ITEM_GROUP).food(ModFoods.apricot);
 	public static final Item.Properties rotten_orange = new Item.Properties().group(Reference.MOD_ITEM_GROUP).food(ModFoods.rotten_orange);
 	public static final Item.Properties rotten_apricot = new Item.Properties().group(Reference.MOD_ITEM_GROUP).food(ModFoods.rotten_apricot);
-	
+	public static final Item.Properties record_colly = new Item.Properties().group(Reference.MOD_ITEM_GROUP).maxStackSize(1);
 }
