+++ src/main/resources/assets/rys/models/item/colly_music_disc.json
@@ -1,0 +1,6 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "rys:item/colly_music_disc"
+  }
+}
